test(desktop): split timeline tool state specs
