+++ apps/desktop/src/composables/__tests__/useTimelineToolState.test.ts
@@ -1,527 +1,30 @@
-import { setupPinia } from "@tracepilot/test-utils";
-import type { ConversationTurn, TurnToolCall } from "@tracepilot/types";
-import { mount } from "@vue/test-utils";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import { defineComponent, nextTick } from "vue";
+import { describe, expect, it } from "vitest";
 import { useSessionDetailStore } from "@/stores/sessionDetail";
-import { useTimelineToolState } from "../useTimelineToolState";
+import { mountTimelineToolState, setupTimelineToolStateTest } from "./useTimelineToolState/setup";
 
-// Test component wrapper that returns the composable's return value
-const TestComponent = defineComponent({
-  setup() {
-    const state = useTimelineToolState();
-    return { ...state };
-  },
-  template: "<div></div>",
-});
+setupTimelineToolStateTest();
 
-describe("useTimelineToolState", () => {
-  beforeEach(() => {
-    setupPinia();
+describe("useTimelineToolState store and loader access", () => {
+  it("provides sessionDetail store", () => {
+    const wrapper = mountTimelineToolState();
+
+    expect(wrapper.vm.store).toBeDefined();
+    expect(wrapper.vm.store).toBe(useSessionDetailStore());
   });
 
-  describe("Store access", () => {
-    it("provides sessionDetail store", () => {
-      const wrapper = mount(TestComponent);
-      expect(wrapper.vm.store).toBeDefined();
-      expect(wrapper.vm.store).toBe(useSessionDetailStore());
-    });
+  it("provides preferences store", () => {
+    const wrapper = mountTimelineToolState();
 
-    it("provides preferences store", () => {
-      const wrapper = mount(TestComponent);
-      expect(wrapper.vm.prefs).toBeDefined();
-    });
+    expect(wrapper.vm.prefs).toBeDefined();
   });
 
-  describe("Tool result loading", () => {
-    it("exposes tool result loader return values", () => {
-      const wrapper = mount(TestComponent);
-      expect(wrapper.vm.fullResults).toBeDefined();
-      expect(wrapper.vm.loadingResults).toBeDefined();
-      expect(wrapper.vm.failedResults).toBeDefined();
-      expect(wrapper.vm.loadFullResult).toBeTypeOf("function");
-      expect(wrapper.vm.retryFullResult).toBeTypeOf("function");
-    });
-  });
-
-  describe("Expansion state", () => {
-    it("provides expandedToolCalls toggle set", () => {
-      const wrapper = mount(TestComponent);
-      const { expandedToolCalls } = wrapper.vm;
-
-      expect(expandedToolCalls.has("test-id")).toBe(false);
-      expandedToolCalls.set.value.add("test-id");
-      expect(expandedToolCalls.has("test-id")).toBe(true);
-      expandedToolCalls.set.value.delete("test-id");
-      expect(expandedToolCalls.has("test-id")).toBe(false);
-    });
-
-    it("provides expandedReasoning toggle set", () => {
-      const wrapper = mount(TestComponent);
-      const { expandedReasoning } = wrapper.vm;
-
-      expect(expandedReasoning.has("reasoning-1")).toBe(false);
-      expandedReasoning.set.value.add("reasoning-1");
-      expect(expandedReasoning.has("reasoning-1")).toBe(true);
-    });
-
-    it("provides expandedOutputs toggle set", () => {
-      const wrapper = mount(TestComponent);
-      const { expandedOutputs } = wrapper.vm;
-
-      expect(expandedOutputs.has("output-1")).toBe(false);
-      expandedOutputs.set.value.add("output-1");
-      expect(expandedOutputs.has("output-1")).toBe(true);
-    });
-
-    it("clears all expansion state with clearAllState", () => {
-      const wrapper = mount(TestComponent);
-      const { expandedToolCalls, expandedReasoning, expandedOutputs, clearAllState } = wrapper.vm;
-
-      expandedToolCalls.set.value.add("tc-1");
-      expandedReasoning.set.value.add("r-1");
-      expandedOutputs.set.value.add("o-1");
-
-      clearAllState();
-
-      expect(expandedToolCalls.has("tc-1")).toBe(false);
-      expect(expandedReasoning.has("r-1")).toBe(false);
-      expect(expandedOutputs.has("o-1")).toBe(false);
-    });
-  });
-
-  describe("Tool call collection", () => {
-    it("computes allToolCalls from store turns", async () => {
-      const wrapper = mount(TestComponent);
-      const store = wrapper.vm.store;
-
-      const mockToolCall1: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        startedAt: "2024-01-01T00:00:00Z",
-        completedAt: "2024-01-01T00:00:01Z",
-        durationMs: 1000,
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      const mockToolCall2: TurnToolCall = {
-        toolName: "read",
-        arguments: '{"path":"file.ts"}',
-        startedAt: "2024-01-01T00:00:02Z",
-        completedAt: "2024-01-01T00:00:03Z",
-        durationMs: 1000,
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-2",
-      };
-
-      const mockTurn1: ConversationTurn = {
-        turnIndex: 0,
-        timestamp: "2024-01-01T00:00:00Z",
-        userMessage: "test message 1",
-        assistantMessages: [],
-        toolCalls: [mockToolCall1],
-        isComplete: true,
-      };
-
-      const mockTurn2: ConversationTurn = {
-        turnIndex: 1,
-        timestamp: "2024-01-01T00:00:05Z",
-        userMessage: undefined,
-        assistantMessages: [],
-        toolCalls: [mockToolCall2],
-        isComplete: true,
-      };
-
-      store.turns = [mockTurn1, mockTurn2];
-      await nextTick();
-
-      expect(wrapper.vm.allToolCalls.length).toBe(2);
-      expect(wrapper.vm.allToolCalls[0].toolCallId).toBe("tc-1");
-      expect(wrapper.vm.allToolCalls[1].toolCallId).toBe("tc-2");
-    });
-  });
-
-  describe("Selected tool management", () => {
-    it("starts with no selection", () => {
-      const wrapper = mount(TestComponent);
-      expect(wrapper.vm.selectedTool).toBe(null);
-    });
-
-    it("selects a tool", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-      expect(wrapper.vm.selectedTool?.toolName).toBe("bash");
-    });
-
-    it("toggles selection when selecting same tool by ID", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool).toBe(null);
-    });
-
-    it("toggles selection when selecting same tool by reference (no ID)", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolName).toBe("bash");
-
-      // Note: Due to Vue's reactivity proxies, the second call may not match
-      // the same object reference, so the toggle may not work as expected
-      // This is a known limitation for tools without IDs
-      wrapper.vm.selectTool(mockTool);
-      // The selection behavior for non-ID tools depends on reference equality
-      // which may not work through Vue proxies, so we just verify it's set
-      expect(wrapper.vm.selectedTool?.toolName).toBeDefined();
-    });
-
-    it("isToolSelected returns true for selected tool by ID", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.isToolSelected(mockTool)).toBe(true);
-
-      // Different object with same ID
-      const sameIdTool: TurnToolCall = {
-        ...mockTool,
-      };
-      expect(wrapper.vm.isToolSelected(sameIdTool)).toBe(true);
-    });
-
-    it("isToolSelected checks tool selection", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      // For tools without IDs, reference equality through Vue proxies may not work reliably
-      // The primary use case is with tools that have IDs
-      expect(wrapper.vm.selectedTool).toBeDefined();
-
-      // Different object without ID should not match (no ID to compare)
-      const differentTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-      };
-      expect(wrapper.vm.isToolSelected(differentTool)).toBe(false);
-    });
-
-    it("isToolSelected returns false for non-selected tool", () => {
-      const wrapper = mount(TestComponent);
-      const toolA: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-      const toolB: TurnToolCall = {
-        toolName: "read",
-        arguments: '{"path":"file.ts"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-2",
-      };
-
-      wrapper.vm.selectTool(toolA);
-      expect(wrapper.vm.isToolSelected(toolB)).toBe(false);
-    });
-
-    it("clears selection with clearSelection", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-
-      wrapper.vm.clearSelection();
-      expect(wrapper.vm.selectedTool).toBe(null);
-    });
-
-    it("clears selection with clearAllState", () => {
-      const wrapper = mount(TestComponent);
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-
-      wrapper.vm.clearAllState();
-      expect(wrapper.vm.selectedTool).toBe(null);
-    });
-
-    it("re-resolves selected tool on data refresh", async () => {
-      const wrapper = mount(TestComponent);
-      const store = wrapper.vm.store;
-
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      const mockTurn: ConversationTurn = {
-        turnIndex: 0,
-        timestamp: "2024-01-01T00:00:00Z",
-        userMessage: "test",
-        assistantMessages: [],
-        toolCalls: [mockTool],
-        isComplete: true,
-      };
-
-      store.turns = [mockTurn];
-      await nextTick();
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-      expect(wrapper.vm.selectedTool?.arguments).toBe('{"command":"ls"}');
-
-      // Simulate data refresh with new object references
-      const newMockTool: TurnToolCall = {
-        ...mockTool,
-        arguments: '{"command":"pwd"}', // Different args but same ID
-      };
-      const newMockTurn: ConversationTurn = {
-        ...mockTurn,
-        toolCalls: [newMockTool],
-      };
-
-      store.turns = [newMockTurn];
-      await nextTick();
-
-      // Should re-resolve to new tool object with same ID
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-      expect(wrapper.vm.selectedTool?.arguments).toBe('{"command":"pwd"}');
-    });
-
-    it("clears selection when selected tool is removed from store", async () => {
-      const wrapper = mount(TestComponent);
-      const store = wrapper.vm.store;
-
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      const mockTurn: ConversationTurn = {
-        turnIndex: 0,
-        timestamp: "2024-01-01T00:00:00Z",
-        userMessage: "test",
-        assistantMessages: [],
-        toolCalls: [mockTool],
-        isComplete: true,
-      };
-
-      store.turns = [mockTurn];
-      await nextTick();
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolCallId).toBe("tc-1");
-
-      // Remove the tool from the store
-      store.turns = [];
-      await nextTick();
-
-      // Selection should be cleared when tool is removed
-      expect(wrapper.vm.selectedTool).toBe(null);
-    });
-
-    it("does not re-resolve if selected tool has no ID", async () => {
-      const wrapper = mount(TestComponent);
-      const store = wrapper.vm.store;
-
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-      };
-
-      const mockTurn: ConversationTurn = {
-        turnIndex: 0,
-        timestamp: "2024-01-01T00:00:00Z",
-        userMessage: "test",
-        assistantMessages: [],
-        toolCalls: [mockTool],
-        isComplete: true,
-      };
-
-      store.turns = [mockTurn];
-      await nextTick();
-
-      wrapper.vm.selectTool(mockTool);
-      expect(wrapper.vm.selectedTool?.toolName).toBe("bash");
-      const originalArgs = wrapper.vm.selectedTool?.arguments;
-
-      // Simulate data refresh
-      const newMockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"pwd"}', // Different args, no ID to match
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-      };
-      const newMockTurn: ConversationTurn = {
-        ...mockTurn,
-        toolCalls: [newMockTool],
-      };
-
-      store.turns = [newMockTurn];
-      await nextTick();
-
-      // Should still reference old arguments (no ID to resolve)
-      expect(wrapper.vm.selectedTool?.arguments).toBe(originalArgs);
-    });
-  });
-
-  describe("Turn ownership check", () => {
-    it("returns undefined when no turnOwnershipCheck provided", () => {
-      const wrapper = mount(TestComponent);
-      expect(wrapper.vm.turnOwnsSelected).toBeUndefined();
-    });
-
-    it("uses custom turnOwnershipCheck function", () => {
-      const turnOwnershipCheck = vi.fn((turn: ConversationTurn, tool: TurnToolCall) => {
-        return turn.toolCalls.some((tc) => tc.toolCallId === tool.toolCallId);
-      });
-
-      const TestComponentWithCheck = defineComponent({
-        setup() {
-          const state = useTimelineToolState({ turnOwnershipCheck });
-          return { ...state };
-        },
-        template: "<div></div>",
-      });
-
-      const wrapper = mount(TestComponentWithCheck);
-      expect(wrapper.vm.turnOwnsSelected).toBeDefined();
-      expect(typeof wrapper.vm.turnOwnsSelected).toBe("function");
-
-      const mockTool: TurnToolCall = {
-        toolName: "bash",
-        arguments: '{"command":"ls"}',
-        success: true,
-        isSubagent: false,
-        isComplete: true,
-        toolCallId: "tc-1",
-      };
-
-      const mockTurn: ConversationTurn = {
-        turnIndex: 0,
-        timestamp: "2024-01-01T00:00:00Z",
-        userMessage: "test",
-        assistantMessages: [],
-        toolCalls: [mockTool],
-        isComplete: true,
-      };
-
-      wrapper.vm.selectTool(mockTool);
-      const result = wrapper.vm.turnOwnsSelected?.(mockTurn);
-
-      expect(result).toBe(true);
-      expect(turnOwnershipCheck).toHaveBeenCalled();
-      const call = turnOwnershipCheck.mock.calls[0];
-      expect(call[0]).toMatchObject(mockTurn);
-      expect(call[1].toolCallId).toBe("tc-1");
-    });
-
-    it("turnOwnsSelected returns false when no tool is selected", () => {
-      const turnOwnershipCheck = vi.fn(() => true);
-
-      const TestComponentWithCheck = defineComponent({
-        setup() {
-          const state = useTimelineToolState({ turnOwnershipCheck });
-          return { ...state };
-        },
-        template: "<div></div>",
-      });
-
-      const wrapper = mount(TestComponentWithCheck);
-
-      const mockTurn: ConversationTurn = {
-        turnIndex: 0,
-        timestamp: "2024-01-01T00:00:00Z",
-        userMessage: "test",
-        assistantMessages: [],
-        toolCalls: [],
-        isComplete: true,
-      };
-
-      const result = wrapper.vm.turnOwnsSelected?.(mockTurn);
-      expect(result).toBe(false);
-      expect(turnOwnershipCheck).not.toHaveBeenCalled();
-    });
+  it("exposes tool result loader return values", () => {
+    const wrapper = mountTimelineToolState();
+
+    expect(wrapper.vm.fullResults).toBeDefined();
+    expect(wrapper.vm.loadingResults).toBeDefined();
+    expect(wrapper.vm.failedResults).toBeDefined();
+    expect(wrapper.vm.loadFullResult).toBeTypeOf("function");
+    expect(wrapper.vm.retryFullResult).toBeTypeOf("function");
   });
 });
